+++ backend/app.py
@@ -1,9 +1,9 @@
 # app.py
 
 from flask import Flask,make_response
-from driver_routes import driver_profile, action_on_request
-from patient_routes import patient_profile, book_ambulance
-from hospital_routes import hospital_details
+from routes.driver_routes import driver_profile, action_on_request
+from routes.patient_routes import patient_profile, book_ambulance
+from routes.hospital_routes import hospital_details
